+++ .eslintrc.js
@@ -1,13 +1,12 @@
 module.exports = {
   extends: ['standard', 'prettier', 'prettier/standard'],
-  plugins: ['prettier', 'standard', 'mocha'],
+  plugins: ['prettier', 'standard', 'import', 'promise'],
   parserOptions: {
     sourceType: 'module'
   },
   env: {
     es6: true,
-    node: true,
-    mocha: true
+    node: true
   },
   rules: {
     'prettier/prettier': ['error', { singleQuote: true, semi: false }],
