+++ package.json
@@ -4,7 +4,7 @@
   "description": "Traverse nested folders and process each of the discovered files.",
   "main": "src/index.js",
   "engines": {
-    "node": ">= 8.12.0"
+    "node": ">= 8.10.0"
   },
   "files": [
     "/package*.json",
@@ -15,8 +15,9 @@
     "eslint-check": "eslint --print-config .eslintrc.js | eslint-config-prettier-check",
     "lint": "eslint .",
     "prettier": "prettier --write '**/*.{js,json,md}'",
-    "test": "find ./test/unit -name '*spec.js' | NODE_ENV=test xargs mocha --require ./test/unit/test_helper.js",
-    "test:coverage": "find ./test/unit -name '*spec.js' | NODE_ENV=test xargs nyc mocha --require ./test/unit/test_helper.js"
+    "test": "find ./test/unit -name '*.test.js' | NODE_ENV=test xargs mocha --require ./test/unit/testHelper.js",
+    "test:coverage": "find ./test/unit -name '*.test.js' | NODE_ENV=test xargs nyc mocha --require ./test/unit/testHelper.js",
+    "test:mutants": "NODE_ENV=test NODE_PATH=. npx stryker run"
   },
   "repository": {
     "type": "git",
@@ -34,13 +35,17 @@
     "traverse"
   ],
   "devDependencies": {
+    "@stryker-mutator/core": "^1.2.0",
+    "@stryker-mutator/javascript-mutator": "^1.2.0",
+    "@stryker-mutator/mocha-framework": "^1.2.0",
+    "@stryker-mutator/mocha-runner": "^1.2.0",
     "ajv": "^6.10.0",
     "chai": "^4.2.0",
     "chai-as-promised": "^7.1.1",
     "eslint": "^5.16.0",
     "eslint-config-prettier": "^4.1.0",
     "eslint-config-standard": "^12.0.0",
-    "eslint-plugin-import": "^2.16.0",
+    "eslint-plugin-import": "^2.17.1",
     "eslint-plugin-mocha": "^5.3.0",
     "eslint-plugin-node": "^8.0.1",
     "eslint-plugin-prettier": "^3.0.1",
@@ -48,9 +53,9 @@
     "eslint-plugin-standard": "^4.0.0",
     "husky": "^1.3.1",
     "lint-staged": "^8.1.5",
-    "mocha": "^6.1.2",
+    "mocha": "^6.1.3",
     "nyc": "^13.3.0",
-    "prettier": "^1.16.4",
+    "prettier": "^1.17.0",
     "proxyquire": "^2.1.0",
     "sinon": "^7.3.1",
     "sinon-chai": "^3.3.0"
