+++ test/.eslintrc.js
@@ -1,6 +1,10 @@
 module.exports = {
+  plugins: ['mocha'],
   rules: {
     'padded-blocks': 0,
     'no-unused-vars': 1
+  },
+  env: {
+    mocha: true
   }
 }
