+++ test/unit/traverse.test.js
